Refinements and monkeypatches organized
