+++ .travis.yml
@@ -4,8 +4,8 @@ rvm:
   - 2.2.5
   - 2.3.1
 env:
-  - TEST_REFINEMENTS=true
-  - TEST_REFINEMENTS=false
+  - TEST_MONKEYPATCHES=true
+  - TEST_MONKEYPATCHES=false
 before_install: 
   - gem install bundler -v 1.12.5
   - curl -sSf https://static.rust-lang.org/rustup.sh | sudo sh -s -- --channel=nightly
