+++ lib/faster_path/version.rb
@@ -1,3 +1,3 @@
 module FasterPath
-  VERSION = "0.0.8"
+  VERSION = "0.0.9"
 end
